# docs/swagger.yaml
basePath: /api
definitions:
...
  core.JSONB:
    additionalProperties: true
    type: object
...
  database.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  github_com_driver005_gateway_internal_setup_intent.CancellationReason:
    enum:
    - abandoned
    - requested_by_customer
    - duplicate
    type: string
    x-enum-varnames:
    - CancellationReasonAbandoned
    - CancellationReasonRequestedByCustomer
    - CancellationReasonDuplicate
  github_com_driver005_gateway_internal_setup_intent.Status:
    enum:
    - requires_payment_method
    - requires_confirmation
    - requires_action
    - processing
    - succeeded
    - failed
    - abandoned
    type: string
    x-enum-varnames:
    - StatusRequiresPaymentMethod
    - StatusRequiresConfirmation
    - StatusRequiresAction
    - StatusProcessing
    - StatusSucceeded
    - StatusFailed
    - StatusAbandoned
...
  intent.SetupIntent:
    properties:
      attach_to_self:
        type: boolean
      cancellation_reason:
        $ref: '#/definitions/github_com_driver005_gateway_internal_setup_intent.CancellationReason'
      client_secret:
        type: string
      created_at:
        type: string
      customer:
        format: uuid
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      description:
        type: string
      flow_directions:
        items:
          type: string
        type: array
      id:
        type: string
      last_setup_error:
        format: uuid
        type: string
      livemode:
        type: boolean
      mandate:
        format: uuid
        type: string
      metadata:
        $ref: '#/definitions/core.JSONB'
      next_action:
        format: uuid
        type: string
      object:
        type: string
      payment_method:
        format: uuid
        type: string
      payment_method_types:
        items:
          type: string
        type: array
      single_use_mandate:
        format: uuid
        type: string
      status:
        $ref: '#/definitions/github_com_driver005_gateway_internal_setup_intent.Status'
      updated_at:
        type: string
      usage:
        $ref: '#/definitions/intent.Usage'
    type: object
  intent.Usage:
    enum:
    - off_session
    - on_session
    type: string
    x-enum-varnames:
    - UsageOffSession
    - UsageOnSession
...
  link.PaymentLink:
    properties:
      active:
        type: boolean
      after_completion:
        format: uuid
        type: string
      allow_promotion_codes:
        type: boolean
      application_fee_amount:
        type: integer
      application_fee_percent:
        type: number
      automatic_tax:
        type: boolean
      billing_address_collection:
        type: string
      consent_collection:
        format: uuid
        type: string
      created_at:
        type: string
      currency:
        type: string
      custom_text:
        format: uuid
        type: string
      customer_creation:
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      id:
        type: string
      line_items:
        items:
          format: uuid
          type: string
        type: array
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      object:
        type: string
      payment_intent_data:
        format: uuid
        type: string
      payment_method_collection:
        type: string
      payment_method_types:
        items:
          type: string
        type: array
      phone_number_collection:
        type: boolean
      shipping_address_collection:
        format: uuid
        type: string
      shipping_options:
        items:
          format: uuid
          type: string
        type: array
      submit_type:
        type: string
      subscription_data:
        format: uuid
        type: string
      tax_id_collection:
        type: boolean
      updated_at:
        type: string
      url:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /v1/payment_links/{id}:
    delete:
      consumes:
      - application/json
      description: Delete PaymentLink by given ID.
      parameters:
      - description: PaymentLink ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete PaymentLink by given ID
      tags:
      - PaymentLink
    get:
      consumes:
      - application/json
      description: Get PaymentLink by given ID or 404 error.
      parameters:
      - description: PaymentLink ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.PaymentLink'
      summary: get PaymentLink by given ID or 404 error.
      tags:
      - PaymentLink
    post:
      consumes:
      - application/json
      description: Update PaymentLink.
      parameters:
      - description: PaymentLink ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.PaymentLink'
      summary: update PaymentLink
      tags:
      - PaymentLink
...
  /v1/setup_intents/{id}:
    delete:
      consumes:
      - application/json
      description: Delete SetupIntent by given ID.
      parameters:
      - description: SetupIntent ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete SetupIntent by given ID
      tags:
      - SetupIntent
    get:
      consumes:
      - application/json
      description: Get SetupIntent by given ID or 404 error.
      parameters:
      - description: SetupIntent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/intent.SetupIntent'
      summary: get SetupIntent by given ID or 404 error.
      tags:
      - SetupIntent
    post:
      consumes:
      - application/json
      description: Update SetupIntent.
      parameters:
      - description: SetupIntent ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/intent.SetupIntent'
      summary: update SetupIntent
      tags:
      - SetupIntent
  /v1/setup_intents/{id}/cancele:
    post:
      consumes:
      - application/json
      description: Cancel SetupIntent.
      parameters:
      - description: SetupIntent ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/intent.SetupIntent'
      summary: cancel SetupIntent
      tags:
      - SetupIntent
  /v1/setup_intents/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Confirm SetupIntent.
      parameters:
      - description: SetupIntent ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/intent.SetupIntent'
      summary: confirm SetupIntent
      tags:
      - SetupIntent
...
schemes:
- http
swagger: "2.0"