# link/handler.go
package link

import (
	"strconv"

	"github.com/driver005/gateway/helper"
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
)

func (h *Handler) SetRoutes(r fiber.Router) {
	r.Get("/payment_links", h.RouteList)
	r.Post("/payment_links", h.RouteCreate)
	r.Get("/payment_links/:id", h.RouteGet)
	r.Post("/payment_links/:id", h.RouteUpdate)
}

// RouteGet func gets PaymentLink by given ID or 404 error.
// @Description Get PaymentLink by given ID or 404 error.
// @Summary get PaymentLink by given ID or 404 error.
// @Tags PaymentLink
// @Accept json
// @Produce json
// @Param id path string true "PaymentLink ID"
// @Success 200 {object} PaymentLink
// @Router /v1/payment_links/{id} [get]
func (h *Handler) RouteGet(context *fiber.Ctx) error {
	Id, err := uuid.Parse(context.Params("id"))
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "parse_error",
		})
	}

	m, err := h.Retrive(context.Context(), Id)
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	return context.Status(fiber.StatusOK).JSON(&m)
}

// RouteList func gets all existing PaymentLinks.
// @Description Get all existing PaymentLinks.
// @Summary get all existing PaymentLinks
// @Tags PaymentLink
// @Accept json
// @Produce json
// @Success 200 {array} PaymentLink
// @Router /v1/payment_links [get]
func (h *Handler) RouteList(context *fiber.Ctx) error {
	page, _ := strconv.Atoi(context.Query("page"))
	if page == 0 {
		page = 1
	}

	pageSize, _ := strconv.Atoi(context.Query("page_size"))
	switch {
	case pageSize > 100:
		pageSize = 100
	case pageSize <= 0:
		pageSize = 10
	}

	offset := (page - 1) * pageSize

	m, n, err := h.List(context.Context(), offset, pageSize)
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	helper.Header(context, context.Request().URI(), *n, pageSize, offset)

	return context.Status(fiber.StatusOK).JSON(&m)
}

// RouteCreate func for creates a new PaymentLink.
// @Description Create a new PaymentLink.
// @Summary create a new PaymentLink
// @Tags PaymentLink
// @Accept json
// @Produce json
// @Success 200 {object} PaymentLink
// @Router /v1/payment_links [post]
func (h *Handler) RouteCreate(context *fiber.Ctx) error {
	var m PaymentLink

	if err := context.BodyParser(&m); err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	r, err := h.Create(context.Context(), &m)
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	return context.Status(fiber.StatusOK).JSON(&r)
}

// RouteUpdate func for updates PaymentLink by given ID.
// @Description Update PaymentLink.
// @Summary update PaymentLink
// @Tags PaymentLink
// @Accept json
// @Produce json
// @Param id body string true "PaymentLink ID"
// @Success 200 {object} PaymentLink
// @Router /v1/payment_links/{id} [post]
func (h *Handler) RouteUpdate(context *fiber.Ctx) error {
	var m PaymentLink

	if err := context.BodyParser(&m); err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	Id, err := uuid.Parse(context.Params("id"))
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "parse_error",
		})
	}

	m.Id = Id

	r, err := h.Update(context.Context(), &m)
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "database_error",
		})
	}

	return context.Status(fiber.StatusOK).JSON(&r)
}

// RouteDelete func for deletes PaymentLink by given ID.
// @Description Delete PaymentLink by given ID.
// @Summary delete PaymentLink by given ID
// @Tags PaymentLink
// @Accept json
// @Produce json
// @Param id body string true "PaymentLink ID"
// @Success 204 {string} status "ok"
// @Router /v1/payment_links/{id} [delete]
func (h *Handler) RouteDelete(context *fiber.Ctx) error {
	Id, err := uuid.Parse(context.Params("id"))
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err.Error(),
			"type":    "parse_error",
		})
	}

	err = h.Delete(context.Context(), Id)
	if err != nil {
		return context.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"message": err,
			"type":    "database_error",
		})
	}

	return context.Status(fiber.StatusNoContent).JSON(nil)
}